pyproject.toml
[tool.poetry]
name = "bot-api"
version = "0.1.0"
description = ""
authors = ["Lana <lan2828@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.1.0"
aiogram = "^3.17.0"
python-dotenv = "^1.0.1"
flask-sqlalchemy = "^3.1.1"
psycopg2-binary = "^2.9.10"
requests = "^2.32.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
